Docker Error Handling (0.5.0)
Update-0.5.0-Docker-Error-Handling

- Test Docker connection

__Test Docker Connection__
Each time we perform a creation or deletion on a test, we recreate the connection to the Docker service. With this I added testing for connection, this will make sure that if the docker instance is shutdown during the running of the application, we will get appropriate errors.


This update appears to be a lot smaller than I originally thought it would be. I do not think that docker errors need to be explicitly handled as most of the time they are good. In future, if we find that any exceptions are not good enough, we can add.